+++ packages/video_player/CHANGELOG.md
@@ -18,3 +18,7 @@
 ## 2.2.1
 
 * Update README
+
+## 2.2.2
+
+* Fix `seekTo` so that it returns when seeking is completed.